[refactor] 모달 중 건물 빌드모달 컴포넌트 분리
